Functions for Discordant Kinship Modeling.

Utilities and functions for discordant kinship modeling (and other sibling-based quasi-experimental designs). Currently, the package contains data restructuring functions and functions for generating biometrically informed data for kin pairs.

Note

The release version is available through CRAN by running install.packages('discord'). The most recent development version is available through GitHub by running devtools::install_github (repo = 'R-Computing-Lab/discord') (make sure devtools is already installed). If you're having trouble with the package, please install the development version. If this doesn't solve your problem, please create a new issue, or email Mason.

References

test

Author

S. Mason Garrison (Wake Forest University) Jonathan Trattner Cermet Ream

Maintainer: S. Mason Garrison <garrissm@wfu.edu>

Examples

library(discord) #Load the package into the current R session.

if (FALSE) {
# Install/update discord with the release version from CRAN.
install.packages('discord')

# Install/update discord with the development version from GitHub
#install.packages('devtools') #Uncomment if `devtools` isn't installed already.
devtools::install_github('R-Computing-Lab/discord')
}