Restructure Data to Determine Kinship Differences

Restructure Data to Determine Kinship Differences

discord_data(
  data,
  outcome,
  predictors,
  id = "extended_id",
  sex = "sex",
  race = "race",
  pair_identifiers,
  demographics = "both"
)

Arguments

data
A data frame.

outcome
A character string containing the outcome variable of interest.

predictors
A character vector containing the column names for predicting the outcome.

id
A unique kinship pair identifier.

sex
A character string for the sex column name.

race
A character string for the race column name.

pair_identifiers
A character vector of length two that contains the variable identifier for each kinship p

demographics
Indicator variable for if the data has the sex and race demographics. If both are present (default, and recommended), value should be "both". Other options include "sex", "race", or "none".

Value

A data frame that

Examples

discord_data(data = uniqueExtendedIDs,
outcome = "flu_2008",
predictors = "edu_2008",
pair_identifiers = c("_s1", "_s2"))

#>     id flu_2008_1 flu_2008_2 flu_2008_diff flu_2008_mean  sex_1  sex_2
#> 1   17          0          0             0           0.0   MALE   MALE
#> 2   29          0          0             0           0.0 FEMALE FEMALE
#> 3   37          1          0             1           0.5 FEMALE FEMALE
#> 4   40          0          0             0           0.0   MALE   MALE
#> 5   58          1          0             1           0.5 FEMALE   MALE
#> 6   61          1          1             0           1.0 FEMALE FEMALE
#> 7   67          0          0             0           0.0 FEMALE   MALE
#> 8   74          0          0             0           0.0   MALE FEMALE
#> 9   83          1          0             1           0.5   MALE   MALE
#> 10  90          1          0             1           0.5   MALE FEMALE
#> 11  94          0          0             0           0.0 FEMALE   MALE
#> 12 107          1          0             1           0.5   MALE   MALE
#> 13 120          1          0             1           0.5 FEMALE   MALE
#> 14 124          0          0             0           0.0   MALE FEMALE
#> 15 137          0          0             0           0.0 FEMALE FEMALE
#> 16 148          1          0             1           0.5 FEMALE   MALE
#> 17 160          0          0             0           0.0 FEMALE FEMALE
#> 18 183          0          0             0           0.0 FEMALE   MALE
#> 19 190          1          0             1           0.5 FEMALE   MALE
#> 20 210          0          0             0           0.0 FEMALE   MALE
#> 21 212          0          0             0           0.0   MALE   MALE
#> 22 219          0          0             0           0.0 FEMALE FEMALE
#> 23 231          0          0             0           0.0   MALE FEMALE
#> 24 238          1          0             1           0.5 FEMALE FEMALE
#> 25 251          1          0             1           0.5   MALE FEMALE
#> 26 254          0          0             0           0.0 FEMALE FEMALE
#> 27 256          0          0             0           0.0   MALE FEMALE
#> 28 260          1          1             0           1.0   MALE   MALE
#> 29 272          1          1             0           1.0   MALE FEMALE
#> 30 274          1          0             1           0.5 FEMALE   MALE
#> 31 281          1          0             1           0.5   MALE FEMALE
#> 32 292          0          0             0           0.0 FEMALE FEMALE
#>                     race_1                  race_2 edu_2008_1 edu_2008_2
#> 1  NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   35.29087  21.279363
#> 2  NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   21.27936  41.447265
#> 3  NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   21.27936  21.279363
#> 4  NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   41.44726  51.403122
#> 5  NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   56.00268  21.279363
#> 6  NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   21.27936  41.447265
#> 7  NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   57.82359  57.823585
#> 8  NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   41.44726  51.403122
#> 9                    BLACK                   BLACK    7.54769   4.272426
#> 10 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   51.40312  51.403122
#> 11 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   21.27936   7.547690
#> 12 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   51.40312  60.385464
#> 13 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   51.40312  51.403122
#> 14 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   41.44726  46.342425
#> 15 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   51.40312  51.403122
#> 16 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   21.27936  21.279363
#> 17 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   41.44726  21.279363
#> 18 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   51.40312  21.279363
#> 19 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   57.82359  57.823585
#> 20 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   51.40312  51.403122
#> 21 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   46.34242  21.279363
#> 22 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   57.82359  60.385464
#> 23 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   21.27936  21.279363
#> 24 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC    7.54769  21.279363
#> 25 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   35.29087  21.279363
#> 26 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   21.27936  21.279363
#> 27 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   21.27936  35.290872
#> 28 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   41.44726  41.447265
#> 29 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   41.44726  51.403122
#> 30 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   41.44726  35.290872
#> 31 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   60.38546  51.403122
#> 32 NON-BLACK, NON-HISPANIC NON-BLACK, NON-HISPANIC   57.82359  21.279363
#>    edu_2008_diff edu_2008_mean
#> 1      14.011509     28.285117
#> 2     -20.167902     31.363314
#> 3       0.000000     21.279363
#> 4      -9.955857     46.425193
#> 5      34.723317     38.641022
#> 6     -20.167902     31.363314
#> 7       0.000000     57.823585
#> 8      -9.955857     46.425193
#> 9       3.275264      5.910058
#> 10      0.000000     51.403122
#> 11     13.731673     14.413527
#> 12     -8.982343     55.894293
#> 13      0.000000     51.403122
#> 14     -4.895160     43.894845
#> 15      0.000000     51.403122
#> 16      0.000000     21.279363
#> 17     20.167902     31.363314
#> 18     30.123758     36.341242
#> 19      0.000000     57.823585
#> 20      0.000000     51.403122
#> 21     25.063062     33.810894
#> 22     -2.561879     59.104525
#> 23      0.000000     21.279363
#> 24    -13.731673     14.413527
#> 25     14.011509     28.285117
#> 26      0.000000     21.279363
#> 27    -14.011509     28.285117
#> 28      0.000000     41.447265
#> 29     -9.955857     46.425193
#> 30      6.156393     38.369068
#> 31      8.982343     55.894293
#> 32     36.544222     39.551474